AG-1660: fix similar genes network chart and table
