+++ libs/agora/genes/src/lib/components/gene-network/gene-network.component.html
@@ -116,7 +116,7 @@
                 these genes are associated with AD.
               </div>
               <div>
-                <i (click)="navigateToSimilarGenes()" class="fa fa-angle-right"></i>
+                <fa-icon (click)="navigateToSimilarGenes()" [icon]="faAngleRight"></fa-icon>
               </div>
             </div>
           </div>
