+++ libs/agora/services/src/index.ts
@@ -1,5 +1,4 @@
 export * from './lib/error.service';
-export * from './lib/gene.service';
 export * from './lib/github.service';
 export * from './lib/helper.service';
 export * from './lib/http-error-interceptor';
